Add support to connect on GCP istance VM through google-sdk in docker env
