+++ README.md
@@ -1,0 +1,34 @@
+# sshmenu fork
+
+sshmenu fork for cluster managemnt
+
+## Install
+
+
+
+## ToDo
+
+Features to add
+- [ ] add new host
+- [ ] delete host
+- [ ] open multiplessh connection
+- [X] use GCP CLI to connect to google cloude istance
+
+## Authors
+
+Contributors names and contact info
+
+Davide Isoardi abacus@linux.it or business mail davide.isoardi@valuepartners.com
+
+## Version History
+
+* 0.1
+    * create annidate menu
+
+## License
+
+This project is licensed under the [GPLv3] License - see the LICENSE.md file for details
+
+## Acknowledgments
+
+Inspiration, code snippets, etc.
